refactor: extract release changelog planning
